+++ app/common/middleware.go
@@ -3,6 +3,8 @@ package common
 import (
     "github.com/go-martini/martini"
     "github.com/martini-contrib/sessions"
+
+    "strconv"
 )
 
 type User interface {
@@ -40,3 +42,21 @@ func UserInject(session sessions.Session, ctx martini.Context) {
 
     ctx.MapTo(u, (*User) (nil))
 }
+
+func CreateData(user User, session sessions.Session) map[string]interface{} {
+    data := map[string]interface{} {
+        "loggedin": strconv.FormatBool(user.IsLoggedIn()),
+        "message": "",
+    }
+    if user.IsLoggedIn() {
+        data["username"] = user.Username()
+    }
+    if session != nil {
+        flashes := session.Flashes("message")
+        if len(flashes) != 0 {
+            data["message"] = flashes[0].(string)
+        }
+    }
+    return data
+}
+
