+++ server.go
@@ -2,6 +2,7 @@ package main
 
 import (
     "github.com/semquery/web/app/common"
+    "github.com/semquery/web/app/routes"
 
     "github.com/go-martini/martini"
     "github.com/martini-contrib/sessions"
@@ -14,41 +15,27 @@ import (
     "gopkg.in/mgo.v2"
 )
 
-var config struct {
-    WebAddr string `json:"web_addr"`
-
-    DBAddr string `json:"db_addr"`
-    DBName string `json:"db_name"`
-    DBUser string `json:"db_user"`
-    DBPass string `json:"db_pass"`
-
-    OAuth2Client_ID string `json:"github_id"`
-    OAuth2Client_Secret string `json:"github_secret"`
-}
-
-var database *mgo.Database
-
 func main() {
     cfg, err := os.Open("config.json")
     if err != nil {
         log.Fatal(err)
     }
     parser := json.NewDecoder(cfg)
-    if err = parser.Decode(&config); err != nil {
-        log.Fatal("Bad json")
+    if err = parser.Decode(common.Config); err != nil {
+        log.Fatal(err)
     }
 
     session, err := mgo.DialWithInfo(&mgo.DialInfo{
-        Addrs: []string{config.DBAddr},
-        Database: config.DBName,
-        Username: config.DBUser,
-        Password: config.DBPass,
+        Addrs: []string{common.Config.DBAddr},
+        Database: common.Config.DBName,
+        Username: common.Config.DBUser,
+        Password: common.Config.DBPass,
     })
     if err != nil {
         log.Fatal(err)
     }
 
-    database = session.DB(config.DBName)
+    common.Database = session.DB(common.Config.DBName)
     log.Print("Database online")
 
     m := martini.Classic()
@@ -58,7 +45,7 @@ func main() {
     }))
     m.Use(common.UserInject)
 
-    RegisterHandlers(m)
+    routes.RegisterRoutes(m)
 
     m.Run()
 }
